requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.64.0

# Database
pymongo>=4.3.0

# LLM Providers
openai>=1.0.0
google-generativeai>=0.3.0

# Text processing
nltk>=3.8.0
spacy>=3.5.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.6.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0

# Utils
python-dotenv>=1.0.0
requests>=2.28.0 